module8/middleware_app/requirements.txt
flask
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp